#230: fixed merge conflicts
