+++ app/src/androidTest/java/com/example/concordia_campus_guide/InstrumentalTests/LocationFragmentInstrumentalTest.java
@@ -10,7 +10,6 @@ import androidx.test.runner.AndroidJUnit4;
 
 import com.example.concordia_campus_guide.activities.MainActivity;
 import com.example.concordia_campus_guide.database.AppDatabase;
-import com.example.concordia_campus_guide.view_models.LocationFragmentViewModel;
 import com.example.concordia_campus_guide.models.Coordinates;
 import com.example.concordia_campus_guide.models.RoomModel;
 import com.google.android.gms.maps.model.PolylineOptions;
