+++ app/src/main/java/com/example/concordia_campus_guide/Helper/UIHelpers/RoomMarkers.java
@@ -1,10 +1,10 @@
-package com.example.concordia_campus_guide.Helper.UIHelpers;
+package com.example.concordia_campus_guide.helper.uiHelpers;
 
 import android.content.Context;
 
 import com.example.concordia_campus_guide.ClassConstants;
-import com.example.concordia_campus_guide.Fragments.LocationFragment.LocationFragmentViewModel;
-import com.example.concordia_campus_guide.Models.RoomModel;
+import com.example.concordia_campus_guide.view_models.LocationFragmentViewModel;
+import com.example.concordia_campus_guide.models.RoomModel;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.BitmapDescriptor;
 import com.google.android.gms.maps.model.LatLng;
